+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Revision history for Z-Redis
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.3.0.0 -- 2021-02-25
 
-* First version. Released on an unsuspecting world.
+* Add `Z.IO.RPC.MessagePack` module.
+* Move some instances to `Z.Data.MessagePack` to make package buildable under constrained memory.
+* Remove `decodeChunks'`.
+* Change `Data.Version.Version`'s instance to use array.